File handling and Data analysis of .tif directory.
Project aims to perform data management of a .tif file directory with sort, move and delete operations using os, shutil and glob modules in tiftask.py. Further it also demonstrates analysis of the files for extraction their geospatial details and metadata utilizing gdal llibrary in raster_meta.py. The co-challenge.zip contains the directory to be operated on.